socialapp/requirements.txt
asgiref==3.6.0
astroid==2.14.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
dill==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
pep8==1.7.1
platformdirs==2.6.2
PyJWT==2.6.0
pylint==2.16.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
wrapt==1.14.1
